+++ app/src/main/java/com/niilopoutanen/rss_feed/activities/ArticleActivity.java
@@ -27,7 +27,7 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior;
 import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.niilopoutanen.rss.Post;
 import com.niilopoutanen.rss_feed.R;
-import com.niilopoutanen.rss_feed.fragments.ArticleView;
+import com.niilopoutanen.rss_feed.fragments.components.ArticleView;
 import com.niilopoutanen.rss_feed.models.Preferences;
 import com.niilopoutanen.rss_feed.utils.PreferencesManager;
 import com.niilopoutanen.rssparser.Callback;
