+++ app/src/main/java/com/niilopoutanen/rss_feed/adapters/DiscoverResultAdapter.java
@@ -5,7 +5,7 @@ import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.niilopoutanen.rss_feed.fragments.SourceItem;
+import com.niilopoutanen.rss_feed.fragments.components.SourceItem;
 import com.niilopoutanen.rss_feed.models.FeedResult;
 
 import java.util.List;
