+++ app/src/main/res/layout/activity_article.xml
@@ -6,7 +6,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context="com.niilopoutanen.rss_feed.activities.ArticleActivity">
 
-    <com.niilopoutanen.rss_feed.fragments.ArticleView
+    <com.niilopoutanen.rss_feed.fragments.components.ArticleView
         android:id="@+id/articleview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
